\name{app_read_csv}
\alias{app_read_csv}
\title{
App Specific Functions
}
\description{
These functions are needed by OpenCPU apps in the package.
}
\usage{
app_read_csv(...)
}
\arguments{
  \item{\dots}{arguments passed to underlying functions.}
}
\details{
\code{app_read_csv} wraps \code{\link{read.csv}}.
}
\author{
Peter Solymos <solymos@ualberta.ca>
}
\keyword{ utils }
